pyproject.toml
[tool.poetry]
name = "su3-distillation"
version = "0.1.0"
description = ""
authors = ["Grant Bradley <g.bradley@fz-juelich.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
h5py = "^3.14.0"
lsqfit = "^13.3.1"
gvar = "^13.1.7"
matplotlib = "^3.10.5"
tqdm = "^4.67.1"
ipykernel = "^6.30.1"
multiprocess = "^0.70.18"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
